Add access level verification to delivery routes

+++ Project/api/routes/delivery.js
@@ -19,18 +19,24 @@ module.exports = (app) => {
                 return;
             }
 
-            var select_all = "SELECT * FROM delivery";
+            if (row.access_lvl >= 1) {
+                var select_all = "SELECT * FROM delivery";
 
-            db.all(select_all, [], (err, rows) => {
-                if (err) {
-                    res.status(400).json({ "error": err.message });
-                    return;
-                }
-                res.status(200).json({
-                    "message": "Successfully retrieved all deliveries",
-                    "data": rows
-                })
-            });
+                db.all(select_all, [], (err, rows) => {
+                    if (err) {
+                        res.status(400).json({ "error": err.message });
+                        return;
+                    }
+                    res.status(200).json({
+                        "message": "Successfully retrieved all deliveries",
+                        "data": rows
+                    })
+                });
+            }
+            else {
+                res.status(401).json({ "error": "Unauthorized" });
+                return;
+            }
         });
     });
 
@@ -52,6 +58,7 @@ module.exports = (app) => {
                 return;
             }
 
+            var user = row;
             var select_one = "SELECT * FROM delivery WHERE id = ?";
 
             db.get(select_one, [req.params.id], (err, row) => {
@@ -63,10 +70,17 @@ module.exports = (app) => {
                     res.status(404).json({ "error": "Delivery not found" });
                     return;
                 }
-                res.status(200).json({
-                    "message": "Successfully retrieved the delivery",
-                    "data": rows
-                })
+
+                if (user.access_lvl >= 1 || user.id === row.employee_id) {
+                    res.status(200).json({
+                        "message": "Successfully retrieved the delivery",
+                        "data": row
+                    })
+                }
+                else {
+                    res.status(401).json({ "error": "Unauthorized" });
+                    return;
+                }
             });
         });
     });
@@ -89,24 +103,30 @@ module.exports = (app) => {
                 return;
             }
 
-            var insert_delivery = "INSERT INTO delivery (employee_id, content, op_date) VALUES (?,?,?)";
-            var params = [req.body.employee_id, req.body.content, req.body.op_date];
+            if (row.access_lvl === 3) {
+                var insert_delivery = "INSERT INTO delivery (employee_id, content, op_date) VALUES (?,?,?)";
+                var params = [req.body.employee_id, req.body.content, req.body.op_date];
 
-            if (req.body.employee_id && req.body.content && req.body.op_date) {
-                let stmt = db.prepare(insert_delivery);
-                stmt.run(params, (err, _) => {
-                    if (err) {
-                        res.status(400).json({ "error": err.message });
-                        return;
-                    }
-                    res.status(200).json({
-                        "message": "Successfully created new delivery",
-                        "data": params
-                    })
-                });
+                if (req.body.employee_id && req.body.content && req.body.op_date) {
+                    let stmt = db.prepare(insert_delivery);
+                    stmt.run(params, (err, _) => {
+                        if (err) {
+                            res.status(400).json({ "error": err.message });
+                            return;
+                        }
+                        res.status(200).json({
+                            "message": "Successfully created new delivery",
+                            "data": params
+                        })
+                    });
+                }
+                else {
+                    res.status(400).json({ "error": "Missing parameters" });
+                }
             }
             else {
-                res.status(400).json({ "error": "Missing parameters" });
+                res.status(401).json({ "error": "Unauthorized" });
+                return;
             }
         });
     });
@@ -129,38 +149,44 @@ module.exports = (app) => {
                 return;
             }
 
-            var find_delivery = "SELECT * FROM delivery WHERE id = ?";
+            if (row.access_lvl === 3) {
+                var find_delivery = "SELECT * FROM delivery WHERE id = ?";
 
-            db.get(find_delivery, [req.params.id], (err, row) => {
-                if (err) {
-                    res.status(400).json({ "error": err.message });
-                    return;
-                }
-                if (!row) {
-                    res.status(404).json({ "error": "Delivery not found" });
-                    return;
-                }
+                db.get(find_delivery, [req.params.id], (err, row) => {
+                    if (err) {
+                        res.status(400).json({ "error": err.message });
+                        return;
+                    }
+                    if (!row) {
+                        res.status(404).json({ "error": "Delivery not found" });
+                        return;
+                    }
 
-                var update_delivery = "UPDATE delivery SET employee_id = ?, content = ?, op_date = ? WHERE id = ?";
-                var params = [req.body.employee_id, req.body.content, req.body.op_date, req.params.id];
+                    var update_delivery = "UPDATE delivery SET employee_id = ?, content = ?, op_date = ? WHERE id = ?";
+                    var params = [req.body.employee_id, req.body.content, req.body.op_date, req.params.id];
 
-                if (req.body.employee_id && req.body.content && req.body.op_date) {
-                    let stmt = db.prepare(update_delivery);
-                    stmt.run(params, (err, _) => {
-                        if (err) {
-                            res.status(400).json({ "error": err.message });
-                            return;
-                        }
-                        res.status(200).json({
-                            "message": "Successfully updated the delivery",
-                            "data": params
-                        })
-                    });
-                }
-                else {
-                    res.status(400).json({ "error": "Missing parameters" });
-                }
-            });
+                    if (req.body.employee_id && req.body.content && req.body.op_date) {
+                        let stmt = db.prepare(update_delivery);
+                        stmt.run(params, (err, _) => {
+                            if (err) {
+                                res.status(400).json({ "error": err.message });
+                                return;
+                            }
+                            res.status(200).json({
+                                "message": "Successfully updated the delivery",
+                                "data": params
+                            })
+                        });
+                    }
+                    else {
+                        res.status(400).json({ "error": "Missing parameters" });
+                    }
+                });
+            }
+            else {
+                res.status(401).json({ "error": "Unauthorized" });
+                return;
+            }
         });
     });
 
@@ -182,31 +208,37 @@ module.exports = (app) => {
                 return;
             }
 
-            var find_delivery = "SELECT * FROM delivery WHERE id = ?";
+            if (row.access_lvl === 3) {
+                var find_delivery = "SELECT * FROM delivery WHERE id = ?";
 
-            db.get(find_delivery, [req.params.id], (err, rows) => {
-                if (err) {
-                    res.status(400).json({ "error": err.message });
-                    return;
-                }
-                if (!rows) {
-                    res.status(404).json({ "error": "Delivery not found" });
-                    return;
-                }
-
-                var delete_delivery = "DELETE FROM delivery WHERE id = ?";
-
-                let stmt = db.prepare(delete_delivery);
-                stmt.run([req.params.id], (err, _) => {
+                db.get(find_delivery, [req.params.id], (err, rows) => {
                     if (err) {
                         res.status(400).json({ "error": err.message });
                         return;
                     }
-                    res.status(200).json({
-                        "message": "Successfully deleted the delivery",
-                    })
+                    if (!rows) {
+                        res.status(404).json({ "error": "Delivery not found" });
+                        return;
+                    }
+
+                    var delete_delivery = "DELETE FROM delivery WHERE id = ?";
+
+                    let stmt = db.prepare(delete_delivery);
+                    stmt.run([req.params.id], (err, _) => {
+                        if (err) {
+                            res.status(400).json({ "error": err.message });
+                            return;
+                        }
+                        res.status(200).json({
+                            "message": "Successfully deleted the delivery",
+                        })
+                    });
                 });
-            });
+            }
+            else {
+                res.status(401).json({ "error": "Unauthorized" });
+                return;
+            }
         });
     });
 };
